pyproject.toml
[tool.poetry]
name = "finlandbustracker"
version = "0.1.0"
description = ""
authors = ["HJacksons <jacksonherberts@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
paho-mqtt = "^2.0.0"
flask = "^3.0.3"
docker = "^7.0.0"
aiohttp = "^3.9.3"
redis = "^5.0.3"
geopy = "^2.4.1"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
